fix: ci stop using actions/checkout@v1
with the https://github.com/ipfs-rust/rust-ipfs/pull/85 there were first
some spurious failures on windows (probably transient gh infra issues)
and later we noticed that the checkout action on lint-rust and
readme-doctest was failing.

I noticed the checkout action on the failing jobs were using an older
version of the failing action than the matrix action which succeeded.
tried to upgrade that and it worked, didn't look further.