add: simple pubsub api
the api is quite horrible with `ipfs::Ipfs::pubsub_*` methods but it
works. `ipfs::Ipfs` provides subscribing to a topic at most once, the
old subscription needs to be unsubscribed with `pubsub_unsubscribe`
before a new one can be started.

message publishing happens via the `Floodsub::publish_any` api which
allows publishing into not subscribed topics.

there are also apis for getting the known peers optionally per topic and
listing the active subscriptions.

subscriptions can be dropped by the user but the subscription will not
be cleaned up until a message comes in... this is perhaps something to
be looked at.