Add pubsub api (#118)
* add: simple pubsub api

the api is quite horrible with `ipfs::Ipfs::pubsub_*` methods but it
works. `ipfs::Ipfs` provides subscribing to a topic at most once, the
old subscription needs to be unsubscribed with `pubsub_unsubscribe`
before a new one can be started.

message publishing happens via the `Floodsub::publish_any` api which
allows publishing into not subscribed topics.

there are also apis for getting the known peers optionally per topic and
listing the active subscriptions.

subscriptions can be dropped by the user but the subscription will not
be cleaned up until a message comes in... this is perhaps something to
be looked at.

* add: simple pubsub tests
* fix: impl unsubscribing via drop
* doc: add missing public doc comment
* doc: add and adjust
* refactor: clippy warnings
* fix: dont unsub on drop after stream ended

this happens only once the sender has been dropped by unsubscribing (or
later via bounded channel becoming full).

* doc: update comments on pubsub
* fix: poll swarm fully

there was an issue where swarm wasn't being polled to the end for which
I hastly created https://github.com/libp2p/rust-libp2p/issues/1516. in
later investigation it turned out that the ipfs::p2p::pubsub was
returning `Poll::Pending` too soon. alternative to the loops would be to
wake the task up where previously `Poll::Pending` was returned.

the issue manifested because of the polling order and the changes made
to the swarm after polling it. the changes created new events which
needed to be dispatched which in turn means the swarm needs to be polled
again or a wakeup be scheduled. adding those wakeups would make the
tasks more co-operative with the same tasks running on the executor.

also reverts the changes made in
290fcfa3ad2dc7cbdca4dfd66511b99a0fd02119 to ipfs::p2p::swarm where
std::task::Waker was recorded and notified on changes. adding the waker
helped in that particular case but not in general.

* fix: remove rare panic in valid situation
* refactor: remove extra logging