fix: move all path matching to v0, box handlers
This patch changes the warp filter boxing strategy so, that:

 1. paths are nested with a "prefix" matcher, if possible
 2. the actual handler for the matched path is boxed
 3. the non implemented but recognized paths are unified

This strategy might be a good compromise between optimizations and
compilation time but such measurements are yet to be done.

Previous boxing strategy ended up being quite unfortunate. Upon adding
the full `/add` implementation, the test cases using GNU binutils linker
or ld 2.33 (ubuntu) started failing failing with a thread overflowing
its stack.

Most of the stack frames were warp filters for matching the path. This
happened because the combine! macro used in http/src/v0.rs boxed
(mostly) each filter, meaning the compiler could never see through any
of the and requiring enough many frames at runtime to cause a stack
overflow.

An alternative of opt-level = 1 turned out to be easy and simple and was
explored in #293 while this solution was being searched for. The
downside is increased compilation time.