feat: initial recursive (un)pinning
needed to turn `recursive: Option<u64>` into custom tri-state
`recursive: Recursive` which has `RecursiveIntent` variant. the "intent"
is used to kickstart the recording of indirect pins for this variant.
in theory we could have crash safety with this marker, but I don't
really think so. it's probably good enough for now.