feat: filesystem backed blockstore
only synchronization needed is the concurrent write blocking
synchronizing per file. this gets rid of the cid, the need to open
anything but now of course everything is more complicated.

I am thinking sure this gets away with the repo level race to
read/complete but yeah, now we have to subscription future systems,
except of course the tokio::sync::broadcast one was much simpler to
implement.