fix: move to semaphore owned permits for locking
this allows using the owned permit and more importantly moving it to the
threadpool thread, which is not necessarily bound to the task lifetime.
there is just one permit per datastore at the moment, but making this
per cid for example shouldn't be too difficult.

having a global lock will be needed later on, not sure how to implement
it though.