fix: panicking while putting a block
See related issue #386, the panic reports were popping up on GHA builds.
The reason why this fixes the issue is not entirely clear, even in
hindsight moving the tx over to the blocking task did seem like the
right thing to do, and it could had enabled other ways to cleanup the
synchronization as well.

However perhaps it is not more explcit that the synchronization key is
created and cleared in the async context.