remove obsolete examples
client1 and client2 were a pair of two peers working together to form up
an example where one client creates an ipld document and the other one
downloads it, reading IPLD projections. the bitswap side is already done
elsewhere and the nature of use having to run two clients, there just
wasn't an easy way around without running the two nodes in the same
process. if that was done, then the example would be the same as
tests.