fix(swarmapi): unit test
this was broken in some upgrade, not the most recent one. it cleverly
only looked like it was doing something but since starting to listen on
an address doesn't do anything but queue up the request, there is
nothing to do for the for-loop and so everything dies down. luckily we
might not need this async unit test here.