fix(swarmapi): fix bug found with fixed test
the test case is rather special one: when the connection is established
there will be a disconnect right away. the unconditional subscription
finishing tripped up the "debug assert" in subscriptions. it might be a
bit wrong to end up with an `Ok(())` value with the test case, but that
is the outcome currently.